Clock service  refactor
