+++ src/test/java/com/example/demox/interfaces/ParkingOperatorControllerTest.java
@@ -1,0 +1,4 @@
+package com.example.demox.interfaces;
+
+public class ParkingOperatorControllerTest {
+}
