+++ src/test/java/com/example/demox/interfaces/ParkingOwnerControllerTest.java
@@ -1,0 +1,4 @@
+package com.example.demox.interfaces;
+
+public class ParkingOwnerControllerTEst {
+}
